+++ src/main/scala/me/piotrjander/cinema/application/exception/BadRequestException.scala
@@ -1,4 +1,4 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.exception
 
 class BadRequestException(msg: String) extends Exception(msg) {
