+++ src/main/scala/me/piotrjander/cinema/application/exception/ValidationException.scala
@@ -1,4 +1,4 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.exception
 
 class ValidationException(msg: String) extends BadRequestException(msg) {
