+++ src/main/scala/me/piotrjander/cinema/application/service/ScreeningService.scala
@@ -1,10 +1,12 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.service
 
 import cats.effect.Async
 import cats.implicits._
-import me.piotrjander.cinema.domain.repository.ScreeningRepository
-import me.piotrjander.cinema.application.ScreeningMessages._
+import me.piotrjander.cinema.application.message.ScreeningMessage._
+import me.piotrjander.cinema.application.validator.DateTimeValidator
+import me.piotrjander.cinema.application.{EntityPayloads, LocalClock}
 import me.piotrjander.cinema.domain.entity.ScreeningId
+import me.piotrjander.cinema.domain.repository.ScreeningRepository
 
 class ScreeningService[F[_]: Async](screeningRepository: ScreeningRepository[F],
                                     localClock: LocalClock[F])
