+++ src/main/scala/me/piotrjander/cinema/application/service/ScreeningServiceApi.scala
@@ -1,0 +1,10 @@
+package me.piotrjander.cinema.application.service
+
+import me.piotrjander.cinema.application.message.ScreeningMessage.{GetRequest, GetResponse, ListRequest, ListResponse}
+
+trait ScreeningServiceApi[F[_]] {
+
+  def list(request: ListRequest): F[ListResponse]
+
+  def get(screeningId: GetRequest): F[Option[GetResponse]]
+}
