+++ src/main/scala/me/piotrjander/cinema/application/validator/DateTimeValidator.scala
@@ -1,9 +1,10 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.validator
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeParseException
 
 import cats.MonadError
+import me.piotrjander.cinema.application.exception.ValidationException
 
 class DateTimeValidator[F[_]](implicit F: MonadError[F, Throwable]) {
