+++ src/main/scala/me/piotrjander/cinema/application/validator/FullNameValidator.scala
@@ -1,13 +1,12 @@
-package me.piotrjander.cinema.application
+package me.piotrjander.cinema.application.validator
 
 import java.util.regex.Pattern
 
 import cats.MonadError
+import me.piotrjander.cinema.application.exception.ValidationException
 import me.piotrjander.cinema.domain.entity.FullName
 
 class FullNameValidator[F[_]](implicit F: MonadError[F, Throwable]) {
-
-  def validationException = new ValidationException("Invalid full name")
 
   def parse(input: String): F[FullName] = {
     val namePartRegex = "\\p{Lu}\\p{Ll}{2,}"
@@ -17,7 +16,7 @@ class FullNameValidator[F[_]](implicit F: MonadError[F, Throwable]) {
     if (matcher.matches()) {
       F.pure(FullName(matcher.group(1), matcher.group(2)))
     } else {
-      F.raiseError(validationException)
+      F.raiseError(new ValidationException())
     }
   }
 }
